requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
absl-py=1.4.0=pypi_0
aiohttp=3.8.3=py38h80987f9_0
aiosignal=1.2.0=pyhd3eb1b0_0
appnope=0.1.3=pyhd8ed1ab_0
asttokens=2.2.1=pyhd8ed1ab_0
async-timeout=4.0.2=py38hca03da5_0
attrs=22.1.0=py38hca03da5_0
autopep8=1.6.0=pyhd3eb1b0_1
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=openblas
blinker=1.4=py38hca03da5_0
brotli=1.0.9=h1a28f6b_7
brotli-bin=1.0.9=h1a28f6b_7
brotlipy=0.7.0=py38h1a28f6b_1002
bzip2=1.0.8=h620ffc9_4
c-ares=1.19.0=h80987f9_0
ca-certificates=2023.01.10=hca03da5_0
cachetools=5.3.0=pypi_0
certifi=2022.12.7=py38hca03da5_0
cffi=1.15.1=py38h80987f9_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py38hca03da5_0
cloudpickle=1.3.0=pypi_0
colorama=0.4.6=pypi_0
contourpy=1.0.5=py38h525c30c_0
cryptography=39.0.1=py38h834c97f_0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.5.1=py38hc377ac9_0
decorator=5.1.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
ffmpeg=4.2.2=h04105a8_0
filelock=3.9.0=py38hca03da5_0
flit-core=3.8.0=py38hca03da5_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h1192e45_0
frozenlist=1.3.3=py38h80987f9_0
future=0.18.3=pypi_0
gettext=0.21.0=h13f89a0_1
giflib=5.2.1=h80987f9_3
gmp=6.2.1=hc377ac9_3
gmpy2=2.1.2=py38h8c48613_0
gnutls=3.6.15=h887c41c_0
google-auth=2.16.3=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.51.3=pypi_0
gym=0.17.2=pypi_0
h5py=3.8.0=pypi_0
icecream=2.1.3=pypi_0
icu=68.1=hc377ac9_0
idna=3.4=py38hca03da5_0
importlib-metadata=6.1.0=pypi_0
importlib_resources=5.2.0=pyhd3eb1b0_1
ipykernel=6.15.0=pyh736e0ef_0
ipython=8.11.0=pyhd1c38e8_0
jedi=0.18.2=pyhd8ed1ab_0
jinja2=3.1.2=py38hca03da5_0
jpeg=9e=h80987f9_1
jupyter_client=7.3.4=pyhd8ed1ab_0
jupyter_core=5.2.0=py38hca03da5_0
kiwisolver=1.4.4=py38h313beb8_0
lame=3.100=h1a28f6b_0
lcms2=2.12=hba8e193_0
lerc=3.0=hc377ac9_0
libbrotlicommon=1.0.9=h1a28f6b_7
libbrotlidec=1.0.9=h1a28f6b_7
libbrotlienc=1.0.9=h1a28f6b_7
libcxx=14.0.6=h848a8c0_0
libdeflate=1.17=h80987f9_0
libffi=3.4.2=hca03da5_6
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libiconv=1.16=h1a28f6b_2
libidn2=2.3.1=h1a28f6b_0
libopenblas=0.3.21=h269037a_0
libopus=1.3=h1a28f6b_1
libpng=1.6.39=h80987f9_0
libprotobuf=3.20.3=h514c7bf_0
libsodium=1.0.18=h27ca646_1
libtasn1=4.16.0=h1a28f6b_0
libtiff=4.5.0=h313beb8_2
libunistring=0.9.10=h1a28f6b_0
libvpx=1.10.0=hc377ac9_0
libwebp=1.2.4=ha3663a8_1
libwebp-base=1.2.4=h80987f9_1
libxml2=2.10.3=h372ba2a_0
llvm-openmp=14.0.6=hc6e5704_0
lz4-c=1.9.4=h313beb8_0
markdown=3.4.3=pypi_0
markupsafe=2.1.1=py38h1a28f6b_0
matplotlib=3.7.1=py38hca03da5_1
matplotlib-base=3.7.1=py38h78102c4_1
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mpc=1.1.0=h8c48613_1
mpfr=4.0.2=h695f6f0_1
mpmath=1.2.1=py38hca03da5_0
multidict=6.0.2=py38h1a28f6b_0
munkres=1.1.4=py_0
ncurses=6.4=h313beb8_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
nettle=3.7.3=h84b5d62_1
networkx=2.8.4=py38hca03da5_1
numpy=1.23.5=py38h1398885_0
numpy-base=1.23.5=py38h90707a3_0
oauthlib=3.2.2=py38hca03da5_0
openh264=1.8.0=h98b2900_0
openssl=1.1.1t=h1a28f6b_0
packaging=23.0=pyhd8ed1ab_0
pandas=1.5.3=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.4.0=py38h313beb8_0
pip=23.0.1=py38hca03da5_0
platformdirs=2.5.2=py38hca03da5_0
prompt-toolkit=3.0.38=pyha770c72_0
prompt_toolkit=3.0.38=hd8ed1ab_0
protobuf=4.22.1=pypi_0
psutil=5.9.0=py38h1a28f6b_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=pypi_0
pycodestyle=2.10.0=py38hca03da5_0
pycparser=2.21=pyhd3eb1b0_0
pygame=2.3.0=pypi_0
pyglet=1.5.0=pypi_0
pygments=2.14.0=pyhd8ed1ab_0
pyjwt=2.4.0=py38hca03da5_0
pyopenssl=23.0.0=py38hca03da5_0
pyparsing=3.0.9=py38hca03da5_0
pysocks=1.7.1=py38hca03da5_0
python=3.8.16=hc0d8a6c_3
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytorch=2.0.0=py3.8_0
pytz=2022.7.1=pypi_0
pyzmq=23.2.0=py38hc377ac9_0
readline=8.2=h1a28f6b_0
requests=2.28.1=py38hca03da5_1
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scipy=1.10.1=pypi_0
seaborn=0.12.2=pypi_0
setuptools=65.6.3=py38hca03da5_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.41.1=h80987f9_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.11.1=py38hca03da5_0
tensorboard=2.10.0=py38hca03da5_0
tensorboard-data-server=0.7.0=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tk=8.6.12=hb8d0fd4_0
toml=0.10.2=pyhd3eb1b0_0
torchaudio=2.0.0=py38_cpu
torchvision=0.15.0=py38_cpu
tornado=6.1=py38hea4295b_1
tqdm=4.65.0=pypi_0
traitlets=5.9.0=pyhd8ed1ab_0
typing_extensions=4.4.0=py38hca03da5_0
urllib3=1.26.14=py38hca03da5_0
wcwidth=0.2.6=pyhd8ed1ab_0
werkzeug=2.2.3=py38hca03da5_0
wheel=0.38.4=py38hca03da5_0
x264=1!152.20180806=h1a28f6b_0
xz=5.2.10=h80987f9_1
yarl=1.8.1=py38h1a28f6b_0
zeromq=4.3.4=hbdafb3b_1
zipp=3.15.0=pypi_0
zlib=1.2.13=h5a0b063_0
zstd=1.5.2=h8574219_0
